+++ core/src/main/kotlin/org/jetbrains/ktfmt/cli/ParsedArgs.kt
@@ -37,7 +37,7 @@ data class ParsedArgs(
 
     /** Return exit code 1 if any formatting changes are detected. */
     val setExitIfChanged: Boolean,
-    /** File name to report when formating code from stdin */
+    /** Path to report for stdin and, when EditorConfig is enabled, use for configuration lookup. */
     val stdinName: String?,
     val editorConfig: Boolean,
     /** Suppress all non-error output. */
@@ -88,7 +88,8 @@ data class ParsedArgs(
         |  --meta-style                      Use 2-space block indenting (default)
         |  --google-style                    Google internal style (2 spaces)
         |  --kotlinlang-style                Kotlin language guidelines style (4 spaces)
-        |  --stdin-name=<name>               Name to report when formatting code from stdin
+        |  --stdin-name=<path>               Path to report for stdin and, when EditorConfig is
+        |                                        enabled, use for configuration lookup
         |  --lines=<lines>                   Line range(s) to format, like 5 or 1:12,14.
         |                                        May be used multiple times.
         |  --offset=<offset>                 Character offset to format, paired with --length.
@@ -129,6 +130,7 @@ data class ParsedArgs(
       var stdinName: String? = null
       var editorConfig = false
       var quiet = false
+      var useExperimentalEngine = false
       val lineRanges = TreeRangeSet.create<Int>()
       val offsets = mutableListOf<Int>()
       val lengths = mutableListOf<Int>()
@@ -158,6 +160,13 @@ data class ParsedArgs(
           arg == "--do-not-remove-unused-imports" -> removeUnusedImports = false
           arg == "--enable-editorconfig" -> editorConfig = true
           arg == "--quiet" -> quiet = true
+          arg == "--experimental-engine" -> {
+            // println, no ceremonies around streams, too experimental
+            System.err.println(
+                "You are using undocumented and unstable feature. We do not recommend doing so",
+            )
+            useExperimentalEngine = true
+          }
           arg.startsWith("--stdin-name=") ->
               stdinName =
                   parseKeyValueArg("--stdin-name", arg)
@@ -254,7 +263,10 @@ data class ParsedArgs(
 
       val parsedArgs = ParsedArgs(
           fileNames,
-          formattingOptions.copy(removeUnusedImports = removeUnusedImports),
+          formattingOptions.copy(
+              removeUnusedImports = removeUnusedImports,
+              experimentalEngine = useExperimentalEngine,
+          ),
           dryRun,
           setExitIfChanged,
           stdinName,